fixed nglTime when rendering with Direct3D (amazing, D3D changes the way time operations work...)
git-svn-id: svn://libnui.net/nui/trunk/nui3@3893 9db30882-7939-4ce0-a36a-45f8a15c01b5

+++ include/nglTime.h
@@ -162,8 +162,8 @@ public:
 private:
   double mValue;
 #ifdef _WIN32_
-  static double mLaunchValue;
-  static int mLaunchDate;
+  static __int64 mLaunchValue;
+  static int32 mLaunchDate;
   static __int64   mTimerFrequency;
   static double GetTime();
   friend class nglTimer;
